dbus-sensors: fixed dbus-sensors crash and can't find sensors issue.
Summary:
The "0005-Utils-support-powerState-for-multi-node-system.patch" does not align with the previous logic. This leads to a crash in the dbus-sensors daemon because it fails to retrieve the host status.

The daemon crash triggers many debug dumps at the same time, using too many resources and causing D-Bus to to hang. As a result, the sensors daemon and PSM daemon also crash,
leading to the error xyz.openbmc_project.Common.Error.InternalFailure.

The commit move patch 0005 to multi-host and add
base on https://gerrit.openbmc.org/c/openbmc/dbus-sensors/+/69161to to fix sensor name not match issue.

X-link: https://github.com/facebookexternal/openbmc.quanta/pull/4530

Reviewed By: williamspatrick

Differential Revision: D67936616

fbshipit-source-id: 40079284dcc218c65ca9b430cb205abd7ffe507e